✨ feat: add max concurrent reconciles to config
This commit adds `maxConcurrentReconciles` to the configuration settings, and sets its default value to 3. The new setting is added to both `chart/questions.yaml` and `chart/values.yaml`. Additionally, this commit modifies `controllers/plug_controller.go` and `controllers/socket_controller.go` to read an environment variable called `MAX_CONCURRENT_RECONCILES` and use its value as the maximum number of concurrent reconciles allowed. If `MAX_CONCURRENT_RECONCILES` is not set or is not a valid integer, then the default value of 3 will be used.