✨ feat: Update integration-operator deployment
This commit updates the environment variable DEBUG_OPERATOR in the integration-operator.yaml deployment file based on the value of .Values.config.debug. If the value of debug is true, then DEBUG_OPERATOR is set to "1", otherwise it's set to "0". Additionally, this commit updates the MAX_CONCURRENT_RECONCILES environment variable with the value of .Values.config.maxConcurrentReconciles.